# By Daniel Cannon
# Via Daniel Cannon
* 'master' of https://github.com/adamstacey/KitchenApplianceCentre:
  More changes to orders
  Updated order service to use new fields
  Fixed order basket template
  Readded product info fields to OrderProduct
  Fixes to the listing template
  Removed inactive controllers
  Removed inactive routes
  Fixed department listing bug caused by change in department path
  More corrections to the checkout section
  More fixes to order/checkout
  Added variant to order product
  Changed order controller to use new entities
  Changed checkout to use new user system
  Fixes to basket
  Added the ability to add a new product to the users basket (WIP)
  Added URL to listing page
  Added product  URL to indexer
  Remove unused routes from old admin bundle
  Fixes to department service to remove references to parentId field

Conflicts:
	src/KAC/SiteBundle/Manager/ProductManager.php

+++ src/KAC/SiteBundle/Command/WarmupIndexCommand.php
@@ -15,11 +15,13 @@ class WarmupIndexCommand extends ContainerAwareCommand
         $this
             ->setName('admin:index:warmup')
             ->setDescription('Load data into Solr index')
+            ->addOption('start', null, InputOption::VALUE_REQUIRED);
         ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+
         gc_enable();
         // Load products
         $em = $this->getContainer()->get('doctrine')->getManager();
@@ -27,12 +29,18 @@ class WarmupIndexCommand extends ContainerAwareCommand
         $productIndexer = new ProductIndexer($this->getContainer()->get('solarium.client.product'), $this->getContainer()->get('doctrine'));
 
         //Clear product index
-        $productIndexer->delete();
+        if($input->getOption('start') === null)
+        {
+            $productIndexer->delete();
+        }
         unset($productIndexer);
 
-        for($i = count($products) - 1; $i >= 0; $i--)
+        $totalProducts = count($products);
+        $startI = $input->getOption('start') == null ? count($products) - 1 : $input->getOption('start');
+
+        for($i = $startI; $i >= 0; $i--)
         {
-            $output->writeln("Writing index for product ".$i." of ".count($products)."...");
+            $output->writeln("Writing index for product ".$i." of ".$totalProducts."...");
             $product = $products[$i];
             unset($products[$i]);
             $productIndexer = new ProductIndexer($this->getContainer()->get('solarium.client.product'), $this->getContainer()->get('doctrine'));
